Fix Win64 build.
BUG=
TEST=

Review URL: https://chromiumcodereview.appspot.com/10414073

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@11627 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/heap.cc
@@ -3326,7 +3326,7 @@ MaybeObject* Heap::AllocateExternalStringFromAscii(
     return Failure::OutOfMemoryException();
   }
 
-  ASSERT(String::IsAscii(resource->data(), length));
+  ASSERT(String::IsAscii(resource->data(), static_cast<int>(length)));
 
   Map* map = external_ascii_string_map();
   Object* result;
